Move Val_t to macos::ffi
